Forgot a check if the account actually still exists :3

+++ src/main/java/mmo/Money/MMOAccount.java
@@ -94,6 +94,10 @@ public class MMOAccount {
     }
     
     public void delete() {
-        MMOMoneyAPI.deleteAccount(this);
+        if (exists()) {
+            MMOMoneyAPI.deleteAccount(this);
+        } else {
+            throw new NullPointerException();
+        }
     }
 }
